fix(backend): add ACCESS_TOKEN_EXPIRATION to .env.example and cast token config to int

+++ backend/config/tokens.php
@@ -10,7 +10,7 @@ return [
     | Default: 15 minutes for security.
     |
     */
-    'access_token_expiration' => env('ACCESS_TOKEN_EXPIRATION', 15),
+    'access_token_expiration' => (int) env('ACCESS_TOKEN_EXPIRATION', 15),
 
     /*
     |--------------------------------------------------------------------------
@@ -21,5 +21,5 @@ return [
     | Default: 10080 minutes (7 days).
     |
     */
-    'refresh_token_expiration' => env('REFRESH_TOKEN_EXPIRATION', 10080),
+    'refresh_token_expiration' => (int) env('REFRESH_TOKEN_EXPIRATION', 10080),
 ];
